drf_api/backend/actual_api/mixins.py
from rest_framework.permissions import IsAdminUser
from .permissions import IsStaffEditorPermission


class StaffEditorPermissionMixin:
    permission_classes = [IsAdminUser, IsStaffEditorPermission]
